Add typed PMA status contracts and extract non-thread CLI groups
Replace phantom dict reads in pma_active/interrupt/reset with typed
dataclasses (PmaActiveResponse, PmaInterruptResponse, PmaResetResponse).
Extract docs, context, binding, and file command groups into dedicated
modules with public helpers. Make hygiene_lock_path public, add shared
resolve_hub_path to pma_control_plane, and replace unbounded SSE
timeout with a bounded httpx.Timeout.
