+++ pyproject.toml
@@ -129,6 +129,10 @@ ignore = ["E501"]
 "src/codex_autorunner/cli.py" = ["B008"]
 "src/codex_autorunner/surfaces/cli/pma_cli.py" = ["B008"]
 "src/codex_autorunner/surfaces/cli/pma_thread_commands.py" = ["B008"]
+"src/codex_autorunner/surfaces/cli/pma_docs_commands.py" = ["B008"]
+"src/codex_autorunner/surfaces/cli/pma_context_commands.py" = ["B008"]
+"src/codex_autorunner/surfaces/cli/pma_binding_commands.py" = ["B008"]
+"src/codex_autorunner/surfaces/cli/pma_file_commands.py" = ["B008"]
 "src/codex_autorunner/surfaces/cli/cli.py" = ["B008"]
 "src/codex_autorunner/surfaces/cli/commands/*.py" = ["B008"]
 "src/codex_autorunner/surfaces/web/routes/terminal_images.py" = ["B008"]
